HMAN-31 moved tests to add verify
