+++ src/integrationTest/java/uk/gov/hmcts/reform/hmc/service/HearingManagementServiceIT.java
@@ -9,7 +9,6 @@ import uk.gov.hmcts.reform.hmc.exceptions.HearingNotFoundException;
 import uk.gov.hmcts.reform.hmc.model.CreateHearingRequest;
 import uk.gov.hmcts.reform.hmc.model.DeleteHearingRequest;
 import uk.gov.hmcts.reform.hmc.model.HearingResponse;
-import uk.gov.hmcts.reform.hmc.model.UpdateHearingRequest;
 import uk.gov.hmcts.reform.hmc.utils.TestingUtil;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -148,114 +147,5 @@ class HearingManagementServiceIT extends BaseTest {
         Exception exception = assertThrows(BadRequestException.class, () -> hearingManagementService
             .deleteHearingRequest(300000000L, request));
         assertEquals("Invalid hearing Id", exception.getMessage());
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenCreateRequestHasPartyDetails() {
-        CreateHearingRequest createHearingRequest = new CreateHearingRequest();
-        createHearingRequest.setRequestDetails(TestingUtil.requestDetails());
-        createHearingRequest.setHearingDetails(TestingUtil.hearingDetails());
-        createHearingRequest.getHearingDetails().setPanelRequirements(TestingUtil.panelRequirements());
-        createHearingRequest.setCaseDetails(TestingUtil.caseDetails());
-        createHearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        createHearingRequest.getPartyDetails().get(0).setOrganisationDetails(TestingUtil.organisationDetails());
-        createHearingRequest.getPartyDetails().get(1).setIndividualDetails(TestingUtil.individualDetails());
-        hearingManagementService.sendRequestToHmi(1L, createHearingRequest);
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenCreateRequestHasOnlyMandatoryFields() {
-        CreateHearingRequest createHearingRequest = new CreateHearingRequest();
-        createHearingRequest.setRequestDetails(TestingUtil.requestDetails());
-        createHearingRequest.setHearingDetails(TestingUtil.hearingDetails());
-        createHearingRequest.getHearingDetails().setPanelRequirements(TestingUtil.panelRequirements());
-        createHearingRequest.setCaseDetails(TestingUtil.caseDetails());
-        hearingManagementService.sendRequestToHmi(1L, createHearingRequest);
-
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenCreateRequestHasNoOrgsDetailsPresent() {
-        CreateHearingRequest createHearingRequest = new CreateHearingRequest();
-        createHearingRequest.setRequestDetails(TestingUtil.requestDetails());
-        createHearingRequest.setHearingDetails(TestingUtil.hearingDetails());
-        createHearingRequest.getHearingDetails().setPanelRequirements(TestingUtil.panelRequirements());
-        createHearingRequest.setCaseDetails(TestingUtil.caseDetails());
-        createHearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        createHearingRequest.getPartyDetails().get(0).setIndividualDetails(TestingUtil.individualDetails());
-        createHearingRequest.getPartyDetails().get(1).setIndividualDetails(TestingUtil.individualDetails());
-        hearingManagementService.sendRequestToHmi(1L, createHearingRequest);
-
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenCreateRequestHasNoIndividualDetailsPresent() {
-        CreateHearingRequest createHearingRequest = new CreateHearingRequest();
-        createHearingRequest.setRequestDetails(TestingUtil.requestDetails());
-        createHearingRequest.setHearingDetails(TestingUtil.hearingDetails());
-        createHearingRequest.getHearingDetails().setPanelRequirements(TestingUtil.panelRequirements());
-        createHearingRequest.setCaseDetails(TestingUtil.caseDetails());
-        createHearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        createHearingRequest.getPartyDetails().get(0).setOrganisationDetails(TestingUtil.organisationDetails());
-        createHearingRequest.getPartyDetails().get(1).setOrganisationDetails(TestingUtil.organisationDetails());
-        hearingManagementService.sendRequestToHmi(1L, createHearingRequest);
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenCreateRequestHasNoRelatedPartyDetailsPresent() {
-        CreateHearingRequest createHearingRequest = new CreateHearingRequest();
-        createHearingRequest.setRequestDetails(TestingUtil.requestDetails());
-        createHearingRequest.setHearingDetails(TestingUtil.hearingDetails());
-        createHearingRequest.getHearingDetails().setPanelRequirements(TestingUtil.panelRequirements());
-        createHearingRequest.setCaseDetails(TestingUtil.caseDetails());
-        createHearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        createHearingRequest.getPartyDetails().get(0).setIndividualDetails(TestingUtil
-                                                                               .individualWithoutRelatedPartyDetails());
-        createHearingRequest.getPartyDetails().get(1).setOrganisationDetails(TestingUtil.organisationDetails());
-        hearingManagementService.sendRequestToHmi(1L, createHearingRequest);
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenUpdateRequestHasPartyDetails() {
-        UpdateHearingRequest hearingRequest = TestingUtil.updateHearingRequest();
-        hearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        hearingRequest.getPartyDetails().get(0).setOrganisationDetails(TestingUtil.organisationDetails());
-        hearingRequest.getPartyDetails().get(1).setIndividualDetails(TestingUtil.individualDetails());
-        hearingManagementService.sendRequestToHmi(1L, hearingRequest);
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenUpdateRequestHasOnlyMandatoryFields() {
-        UpdateHearingRequest hearingRequest = TestingUtil.updateHearingRequest();
-        hearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        hearingManagementService.sendRequestToHmi(1L, hearingRequest);
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenUpdateRequestHasNoOrgsDetailsPresent() {
-        UpdateHearingRequest hearingRequest = TestingUtil.updateHearingRequest();
-        hearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        hearingRequest.getPartyDetails().get(0).setIndividualDetails(TestingUtil.individualDetails());
-        hearingRequest.getPartyDetails().get(1).setIndividualDetails(TestingUtil.individualDetails());
-        hearingManagementService.sendRequestToHmi(1L, hearingRequest);
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenUpdateRequestHasNoIndividualDetailsPresent() {
-        UpdateHearingRequest hearingRequest = TestingUtil.updateHearingRequest();
-        hearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        hearingRequest.getPartyDetails().get(0).setOrganisationDetails(TestingUtil.organisationDetails());
-        hearingRequest.getPartyDetails().get(1).setOrganisationDetails(TestingUtil.organisationDetails());
-        hearingManagementService.sendRequestToHmi(1L, hearingRequest);
-    }
-
-    @Test
-    void shouldSuccessfullyMapToHmiFormatWhenUpdateRequestHasNoRelatedPartyDetailsPresent() {
-        UpdateHearingRequest hearingRequest = TestingUtil.updateHearingRequest();
-        hearingRequest.setPartyDetails(TestingUtil.partyDetails());
-        hearingRequest.getPartyDetails().get(0).setOrganisationDetails(TestingUtil.organisationDetails());
-        hearingRequest.getPartyDetails().get(1).setIndividualDetails(TestingUtil
-                                                                         .individualWithoutRelatedPartyDetails());
-        hearingManagementService.sendRequestToHmi(1L, hearingRequest);
     }
 }
